Add unit tests for system size calculations
branch:
